Moves ledger notification into seperate files
Resolves #11876

Auditors:

Test Plan:
